mbw lib: get mbw-monitor code commit e5609cfaed68ef53441b1ad8010181157794b05c, June 3 2024

+++ pkg/mbw/utils/pci/pciutils.go
@@ -166,11 +166,11 @@ func ReadLong(dev *PCIDev, addr int32) uint32 {
 }
 
 func ReadSMNApp(dev *PCIDev, addr uint32) uint32 {
-	WriteLong(dev, IOHC_NB_SMN_INDEX_0_REG, addr)
+	WriteLong(dev, IOHC_NB_SMN_INDEX_0_REG, uint32(addr))
 	return ReadLong(dev, IOHC_NB_SMN_DATA_0_REG)
 }
 
 func WriteSMNApp(dev *PCIDev, addr, data uint32) {
-	WriteLong(dev, IOHC_NB_SMN_INDEX_0_REG, addr)
+	WriteLong(dev, IOHC_NB_SMN_INDEX_0_REG, uint32(addr))
 	WriteLong(dev, IOHC_NB_SMN_DATA_0_REG, data)
 }
